added token, login

+++ controllers/cards.js
@@ -1,32 +1,67 @@
+const escape = require('escape-html');
+
 const Cards = require('../models/card');
-const { handleEmptyData } = require('./services/dataHandlers');
-const { checkIdValiness } = require('./services/checkIdValiness');
+const { VALIDATION_ERROR } = require('../models/services/validateService');
+const { checkIdValidness } = require('./services/checkIdValidness');
 const { handleNotFound } = require('./services/handleNotFound');
 
 module.exports.readCards = (req, res) => {
   Cards.find({})
-    .then((cards) => handleEmptyData(cards, res))
-    .then((cards) => res.send({ data: cards }))
+    .then((cards) => {
+      if (!cards) {
+        handleNotFound(res);
+      } else {
+        res.send({ data: cards });
+      }
+    })
     .catch((err) => res.status(500).send({ message: `Произошла ошибка, ${err}` }));
 };
 
 module.exports.createCard = (req, res) => {
   const { name, link } = req.body;
 
-  Cards.create({ name, link, owner: req.user._id })
+  Cards.create({
+    name: escape(name),
+    link: escape(link),
+    owner: req.user._id,
+  })
     .then((card) => res.send({ data: card }))
-    .catch((err) => res.status(500).send({ message: `Произошла ошибка, ${err}` }));
+    .catch((err) => {
+      const isValidationError = err.name.startsWith(VALIDATION_ERROR);
+      if (isValidationError) {
+        res.status(400).send({ message: `Произошла ошибка, ${err}` });
+      } else {
+        res.status(500).send({ message: `Произошла ошибка, ${err}` });
+      }
+    });
 };
 
 module.exports.deleteCard = async (req, res) => {
   const { cardId } = req.params;
-  const isIdValid = await checkIdValiness(Cards, cardId);
+  let isIdValid = checkIdValidness(cardId);
 
   if (isIdValid) {
-    Cards.findByIdAndRemove(cardId)
-      .then((card) => handleEmptyData(card, res))
-      .then(() => res.send({ message: 'данные обновлены' }))
+    const { _id } = req.user;
+    let isUserCardOwner = false;
+    await Cards.findById(cardId)
+      .then((card) => {
+        if (!card) {
+          isIdValid = false;
+        } else {
+          isUserCardOwner = String(_id) === String(card.owner);
+        }
+      })
       .catch((err) => res.status(500).send({ message: `Произошла ошибка, ${err}` }));
+
+    if (isIdValid && isUserCardOwner) {
+      Cards.findByIdAndRemove(cardId)
+        .then(() => res.send({ message: 'данные обновлены' }))
+        .catch((err) => res.status(500).send({ message: `Произошла ошибка, ${err}` }));
+    } else if (!isIdValid) {
+      handleNotFound(res);
+    } else if (!isUserCardOwner) {
+      res.status(403).send({ message: 'действие недоступно пользователю' });
+    }
   } else {
     handleNotFound(res);
   }
@@ -34,7 +69,7 @@ module.exports.deleteCard = async (req, res) => {
 
 module.exports.likeCard = async (req, res) => {
   const { cardId } = req.params;
-  const isIdValid = await checkIdValiness(Cards, cardId);
+  const isIdValid = await checkIdValidness(cardId);
 
   if (isIdValid) {
     Cards.findByIdAndUpdate(
@@ -42,8 +77,13 @@ module.exports.likeCard = async (req, res) => {
       { $addToSet: { likes: req.user._id } },
       { new: true },
     )
-      .then((card) => handleEmptyData(card, res))
-      .then((card) => res.send({ data: card }))
+      .then((card) => {
+        if (!card) {
+          handleNotFound(res);
+        } else {
+          res.send({ data: card });
+        }
+      })
       .catch((err) => res.status(500).send({ message: `Произошла ошибка, ${err}` }));
   } else {
     handleNotFound(res);
@@ -52,7 +92,7 @@ module.exports.likeCard = async (req, res) => {
 
 module.exports.dislikeCard = async (req, res) => {
   const { cardId } = req.params;
-  const isIdValid = await checkIdValiness(Cards, cardId);
+  const isIdValid = await checkIdValidness(cardId);
 
   if (isIdValid) {
     Cards.findByIdAndUpdate(
@@ -60,8 +100,13 @@ module.exports.dislikeCard = async (req, res) => {
       { $pull: { likes: req.user._id } },
       { new: true },
     )
-      .then((card) => handleEmptyData(card, res))
-      .then((card) => res.send({ data: card }))
+      .then((card) => {
+        if (!card) {
+          handleNotFound(res);
+        } else {
+          res.send({ data: card });
+        }
+      })
       .catch((err) => res.status(500).send({ message: `Произошла ошибка, ${err}` }));
   } else {
     handleNotFound(res);
